+++ Environments/SharePoint-AllVersions/dsc/ConfigureSQLVM.ps1
@@ -9,7 +9,7 @@ configuration ConfigureSQLVM
         [Parameter(Mandatory)] [System.Management.Automation.PSCredential]$SPSetupCreds
     )
 
-    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, xActiveDirectory, SqlServerDsc, xPSDesiredStateConfiguration
+    Import-DscResource -ModuleName ComputerManagementDsc, NetworkingDsc, ActiveDirectoryDsc, SqlServerDsc, xPSDesiredStateConfiguration
 
     WaitForSqlSetup
     [String] $DomainNetbiosName = (Get-NetBIOSName -DomainFQDN $DomainFQDN)
@@ -19,8 +19,6 @@ configuration ConfigureSQLVM
     [PSCredential] $SPSetupCredsQualified = New-Object PSCredential ("${DomainNetbiosName}\$($SPSetupCreds.UserName)", $SPSetupCreds.Password)
     [PSCredential] $SQLCredsQualified = New-Object PSCredential ("${DomainNetbiosName}\$($SqlSvcCreds.UserName)", $SqlSvcCreds.Password)
     $ComputerName = Get-Content env:computername
-    [Int] $RetryCount = 30
-    [Int] $RetryIntervalSec = 30
 
     Node localhost
     {
@@ -40,13 +38,14 @@ configuration ConfigureSQLVM
         #**********************************************************
         # Join AD forest
         #**********************************************************
-        xWaitForADDomain DscForestWait
+        WaitForADDomain DscForestWait
         {
-            DomainName           = $DomainFQDN
-            DomainUserCredential = $DomainAdminCredsQualified
-            RetryCount           = $RetryCount
-            RetryIntervalSec     = $RetryIntervalSec
-            #DependsOn            = "[DnsServerAddress]DnsServerAddress"
+            DomainName              = $DomainFQDN
+            WaitTimeout             = 600
+            RestartCount            = 2
+            WaitForValidCredentials = $True
+            Credential              = $DomainAdminCredsQualified
+            DependsOn               = "[DnsServerAddress]DnsServerAddress"
         }
 
         Computer DomainJoin
@@ -54,57 +53,64 @@ configuration ConfigureSQLVM
             Name = $ComputerName
             DomainName = $DomainFQDN
             Credential = $DomainAdminCredsQualified
-            DependsOn = "[xWaitForADDomain]DscForestWait"
+            DependsOn = "[WaitForADDomain]DscForestWait"
+        }
+
+        PendingReboot RebootAfterJoinedDomain
+        {
+            Name             = "RebootAfterJoinedDomain"
+            SkipCcmClientSDK = $true
+            DependsOn        = "[Computer]DomainJoin"
         }
 
         #**********************************************************
         # Create accounts and configure SQL Server
         #**********************************************************
-        xADUser CreateSqlSvcAccount
+        ADUser CreateSqlSvcAccount
         {
-            DomainAdministratorCredential = $DomainAdminCredsQualified
             DomainName = $DomainFQDN
             UserName = $SqlSvcCreds.UserName
             Password = $SQLCredsQualified
             PasswordNeverExpires = $true
             Ensure = "Present"
-            DependsOn = "[Computer]DomainJoin"
+            PsDscRunAsCredential = $DomainAdminCredsQualified
+            DependsOn = "[PendingReboot]RebootAfterJoinedDomain"
         }
 
-        xADServicePrincipalName UpdateSqlSPN1
+        ADServicePrincipalName UpdateSqlSPN1
         {
             Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName.$($DomainFQDN):1433"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN2
+        ADServicePrincipalName UpdateSqlSPN2
         {
             Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName.$DomainFQDN"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN3
+        ADServicePrincipalName UpdateSqlSPN3
         {
             Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$($ComputerName):1433"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
-        xADServicePrincipalName UpdateSqlSPN4
+        ADServicePrincipalName UpdateSqlSPN4
         {
             Ensure               = "Present"
             ServicePrincipalName = "MSSQLSvc/$ComputerName"
             Account              = $SqlSvcCreds.UserName
             PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[xADUser]CreateSqlSvcAccount"
+            DependsOn            = "[ADUser]CreateSqlSvcAccount"
         }
 
         SqlServiceAccount SetSqlInstanceServiceAccount
@@ -114,18 +120,18 @@ configuration ConfigureSQLVM
             ServiceType    = "DatabaseEngine"
             ServiceAccount = $SQLCredsQualified
             RestartService = $true
-            DependsOn      = "[xADServicePrincipalName]UpdateSqlSPN1", "[xADServicePrincipalName]UpdateSqlSPN2", "[xADServicePrincipalName]UpdateSqlSPN3", "[xADServicePrincipalName]UpdateSqlSPN4"
+            DependsOn      = "[ADServicePrincipalName]UpdateSqlSPN1", "[ADServicePrincipalName]UpdateSqlSPN2", "[ADServicePrincipalName]UpdateSqlSPN3", "[ADServicePrincipalName]UpdateSqlSPN4"
         }
 
-        xADUser CreateSPSetupAccount
+        ADUser CreateSPSetupAccount
         {
-            DomainAdministratorCredential = $DomainAdminCredsQualified
             DomainName = $DomainFQDN
             UserName = $SPSetupCreds.UserName
             Password = $SPSetupCredsQualified
             PasswordNeverExpires = $true
             Ensure = "Present"
-            DependsOn = "[Computer]DomainJoin"
+            PsDscRunAsCredential = $DomainAdminCredsQualified
+            DependsOn = "[PendingReboot]RebootAfterJoinedDomain"
         }
 
         SqlLogin AddDomainAdminLogin
@@ -135,7 +141,7 @@ configuration ConfigureSQLVM
             ServerName = $ComputerName
             InstanceName = "MSSQLSERVER"
             LoginType = "WindowsUser"
-            DependsOn = "[Computer]DomainJoin"
+            DependsOn = "[PendingReboot]RebootAfterJoinedDomain"
         }
 
         SqlLogin AddSPSetupLogin
@@ -145,7 +151,7 @@ configuration ConfigureSQLVM
             ServerName = $ComputerName
             InstanceName = "MSSQLSERVER"
             LoginType = "WindowsUser"
-            DependsOn = "[xADUser]CreateSPSetupAccount"
+            DependsOn = "[ADUser]CreateSPSetupAccount"
         }
 
         SqlRole GrantSQLRoleSysadmin
@@ -183,7 +189,7 @@ configuration ConfigureSQLVM
             ServerName   = $ComputerName
             InstanceName = "MSSQLSERVER"
             MaxDop       = 1
-            DependsOn    = "[Computer]DomainJoin"
+            DependsOn    = "[PendingReboot]RebootAfterJoinedDomain"
         }
 
         # Open port on the firewall when everything is ready, as SharePoint DSC is testing it to start creation of the farm
@@ -257,7 +263,8 @@ $SPSetupCreds = Get-Credential -Credential "spsetup"
 $DNSServer = "10.0.1.4"
 $DomainFQDN = "contoso.local"
 
-ConfigureSQLVM -DNSServer $DNSServer -DomainFQDN $DomainFQDN -DomainAdminCreds $DomainAdminCreds -SqlSvcCreds $SqlSvcCreds -SPSetupCreds $SPSetupCreds -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath "C:\Data\\output"
-Start-DscConfiguration -Path "C:\Data\output" -Wait -Verbose -Force
+$outputPath = "C:\Packages\Plugins\Microsoft.Powershell.DSC\2.80.0.3\DSCWork\ConfigureSQLVM.0\ConfigureSQLVM"
+ConfigureSQLVM -DNSServer $DNSServer -DomainFQDN $DomainFQDN -DomainAdminCreds $DomainAdminCreds -SqlSvcCreds $SqlSvcCreds -SPSetupCreds $SPSetupCreds -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath $outputPath
+Start-DscConfiguration -Path $outputPath -Wait -Verbose -Force
 
 #>